Changed the way the animations work. No longer need the animator system or the FPS variable in the SpriteSheet component.
Refactor and enhance invader control and system updates

Refactored `Sys_InvaderControl` to improve invader loading and beat handling. Added new methods for invader initialization and beat updates. Updated various system methods to use `const auto&` for component retrieval. Enhanced bullet control, collision handling, and rendering logic. Adjusted game state management and event handling for better performance and clarity.

+++ State.h
@@ -32,5 +32,5 @@ protected:
 	sf::View m_fullWindowView;
 	bool m_transparent;
 	sf::Font m_font;
-	float m_fontSize = 100.f;
+	unsigned int m_fontSize = 100;
 };
